requirements.txt
llama-index>=0.9.8
openai>=1.12.0
python-dotenv>=1.0.0
click>=8.1.7
PyYAML>=6.0.1
python-magic>=0.4.27
tqdm>=4.66.1
python-docx>=1.0.1
PyPDF2>=3.0.1
rich>=13.7.0 